telemetry-server/requirements.txt
fastapi
uvicorn
pymavlink
websockets
google-cloud-pubsub
